requirements.txt
blinker==1.4
click==8.0.3
colorama==0.4.4
dataclasses==0.8
Flask==2.0.2
Flask-Mail==0.9.1
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
typing-extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
